Add MasslessChargedParticleSingular problem
Add a second massless charged particle problem that uses a "singular"
(one-component) magnetic vector potential with A₂ = 0, chosen as a gauge
transformation that reproduces the same magnetic field and dynamics as the
standard problem. This one-form is the form required by the degenerate
variational integrator (DVI) method of GeometricIntegrators.

Common functionality is factored into massless_charged_particle_common.jl,
included by both massless_charged_particle.jl and the new
massless_charged_particle_singular.jl (mirroring the LotkaVolterra2dSingular
pattern). Each module defines only the gauge-specific primitives (A₁, A₂,
dϑ derivatives, B) and includes the common file.

- Refactor the Makie plot extension to serve both modules via shared helpers.
- Add tests mirroring the standard problem plus same-B/same-v cross-checks.
- Extend the documentation with a singular vector potential section.
