+++ src/GeometricProblems.jl
@@ -23,6 +23,7 @@ module GeometricProblems
     include("lotka_volterra_4d.jl")
     include("lotka_volterra_4d_lagrangian.jl")
     include("massless_charged_particle.jl")
+    include("massless_charged_particle_singular.jl")
     include("coupled_harmonic_oscillator.jl")
     include("mathews_lakshmanan_oscillator.jl")
     include("morse_oscillator.jl")
